+++ shared/params/config.go
@@ -316,6 +316,7 @@ func E2ETestConfig() *BeaconChainConfig {
 	e2eConfig.SecondsPerSlot = 8
 	e2eConfig.SecondsPerETH1Block = 2
 	e2eConfig.Eth1FollowDistance = 4
+	e2eConfig.PersistentCommitteePeriod = 4
 	return e2eConfig
 }
